move functions into utils file
