+++ uclasm/matching/search/search_utils.py
@@ -1,0 +1,57 @@
+"""Utility functions and classes for search"""
+import numpy as np
+
+class State:
+    """A state for the greedy search algorithm.
+    Attributes
+    ----------
+    matching : tuple
+        Tuple representation of the current matching.
+    cost: float
+        Estimated cost of the matching.
+    """
+    def __init__(self):
+        self.matching = None
+        self.cost = float("inf")
+
+    def __lt__(self, other):
+        return self.cost < other.cost
+
+    def __str__(self):
+        return str(self.matching) + ": " + str(self.cost)
+
+def tuple_from_dict(dict):
+    """Turns a dict into a representative sorted tuple of 2-tuples.
+    Parameters
+    ----------
+    dict: dict
+        A dictionary.
+    Returns
+    -------
+    tuple
+        A unique tuple representation of the dictionary as a sequence of
+        2-tuples of key-value pairs, sorted by the dictionary key.
+    """
+    return tuple(sorted(dict.items()))
+
+def dict_from_tuple(tuple):
+    """Turns a tuple of 2-tuples into a dict.
+    Parameters
+    ----------
+    tuple: tuple
+        A unique tuple representation of the dictionary as a sequence of
+        2-tuples of key-value pairs. Keys must be unique.
+    Returns
+    -------
+    dict
+        The equivalent dictionary.
+    """
+    return dict(tuple)
+
+def set_fixed_costs(fixed_costs, matching):
+    """Set fixed costs to float("inf") to enforce the given matching."""
+    mask = np.zeros(fixed_costs.shape, dtype=np.bool)
+    mask[[pair[0] for pair in matching],:] = True
+    mask[:,[pair[1] for pair in matching]] = True
+    mask[tuple(np.array(matching).T)] = False
+    fixed_costs[mask] = float("inf")
